feat(graphql): implement features to support dashboard automated analysis card (#312)
* feat(graphql): implement features to support dashboard automated analysis card

* disable JVM ID node filtering, populate archived recordings query with data

* populate aggregate size

* refactor

* refactor to split out class
